+++ tests/test_registry.py
@@ -39,8 +39,26 @@ def test_injector_container_inject_props():
     registry = InjectorRegistry()
     registry.register_injectable(Heading, use_props=True)
     container = registry.create_injectable_container()
-    heading: Heading = container.inject(Heading, first_name='Injected Name')
+    heading: Heading = container.inject(
+        Heading, first_name='Injected Name'
+    )
     assert 'Injected Name' == heading.first_name
+
+
+def test_injector_container_inject_name_prop():
+    # Injecting a prop of ``name`` collides with the usage of
+    # container.get(iface, name=)
+    @dataclass
+    class Heading:
+        name: str = 'Default Name'
+
+    registry = InjectorRegistry()
+    registry.register_injectable(Heading, use_props=True)
+    container = registry.create_injectable_container()
+    heading: Heading = container.inject(
+        Heading, name='Injected Name'
+    )
+    assert 'Injected Name' == heading.name
 
 
 def test_injector_registry_scan_caller():
@@ -80,4 +98,3 @@ def test_injector_target_none():
     container = registry.create_injectable_container()
     heading: Heading = container.get(Heading)
     assert 'Default Name' == heading.name
-