+++ CHANGES.md
@@ -1,5 +1,9 @@
 # Changes
 
+## Unreleased
+
+- Change `InjectorContainer.inject` to accept `cget_props` as a way of avoiding clashes between props and underlying `.get()` args
+
 ## 0.3.0
 
 - Allow register_injectable to only supply `for_` and `target` then defaults to it